app/requirements.txt
flask
mysql-connector-python 
numpy
pandas
keras
matplotlib
tensorflow
librosa
path.py
pillow